requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.12.23
requests>=2.31.0
python-dotenv==1.0.0 
brotli>=1.0.9 
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0 